+++ src/app/core/abstractions/ad-lib-piece.service.ts
@@ -1,9 +1,7 @@
 import {Observable} from 'rxjs';
 import {Identifier} from '../models/identifier';
 
-export interface AdLibPieceServiceInterface {
-
-  fetchAdLibPieceIdentifiers(rundownId: string): Observable<Identifier[]>
-
-  executeAdLibPiece(rundownId: string, adLibIdentifier: Identifier): Observable<void>
+export abstract class AdLibPieceService {
+  public abstract fetchAdLibPieceIdentifiers(rundownId: string): Observable<Identifier[]>
+  public abstract executeAdLibPiece(rundownId: string, adLibIdentifier: Identifier): Observable<void>
 }
