+++ src/app/core/core.module.ts
@@ -5,7 +5,8 @@ import { HttpErrorService } from './services/http-error.service'
 import { HttpClientModule } from '@angular/common/http'
 import { HttpBasicRundownService } from './services/http-basic-rundown.service'
 import { HttpRundownService } from './services/http-rundown.service'
-import { AdLibPieceService } from './services/ad-lib-piece.service'
+import { AdLibPieceService } from './abstractions/ad-lib-piece.service'
+import { HttpAdLibPieceService } from './services/http-ad-lib-piece.service'
 import { ConnectionStatusObserver } from './services/connection-status-observer.service'
 import { RundownEventObserver } from './services/rundown-event-observer.service'
 import { RundownStateService } from './services/rundown-state.service'
@@ -27,7 +28,7 @@ import { ZodRundownEventParser } from './parsers/zod-rundown-event-parser.servic
     HttpClientModule,
     HttpBasicRundownService,
     HttpRundownService,
-    AdLibPieceService,
+    { provide: AdLibPieceService, useClass: HttpAdLibPieceService },
     ConnectionStatusObserver,
     RundownEventObserver,
     { provide: BasicRundownService, useClass: HttpBasicRundownService },
